Renamed SettingSwitch to Switch.
